# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API UHS Asbl",
        "description": "API REST complète avec gestion des activites de UHS Asbl",
        "contact": {
            "name": "Support API",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-uhs.uhs-rdc.org",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/point-eaux.store": {
            "post": {
                "tags": [
                    "Points d’eau"
                ],
                "summary": "Créer un point d’eau",
                "operationId": "654581a9b53fb73d54ba41d0bd72d4e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lat",
                                    "long",
                                    "status",
                                    "numero_compteur"
                                ],
                                "properties": {
                                    "lat": {
                                        "type": "string",
                                        "example": "-1.6789"
                                    },
                                    "long": {
                                        "type": "string",
                                        "example": "29.2345"
                                    },
                                    "numero_compteur": {
                                        "type": "string",
                                        "example": "COMP-001"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Actif"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Point d’eau créé avec succès"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/point-eaux.update/{id}": {
            "put": {
                "tags": [
                    "Points d’eau"
                ],
                "summary": "Mettre à jour un point d’eau",
                "operationId": "468b286027dfd3599955d452e2884b3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du point d’eau",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lat": {
                                        "type": "string",
                                        "example": "-1.6789"
                                    },
                                    "long": {
                                        "type": "string",
                                        "example": "29.2345"
                                    },
                                    "numero_compteur": {
                                        "type": "string",
                                        "example": "COMP-001"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "Actif"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Point d’eau mis à jour avec succès"
                    },
                    "404": {
                        "description": "Point d’eau non trouvé"
                    }
                }
            }
        },
        "/api/point-eaux.delete/{id}": {
            "delete": {
                "tags": [
                    "Points d’eau"
                ],
                "summary": "Supprimer un point d’eau",
                "operationId": "1ff1a61a7dd52ae4d07541151365b4ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du point d’eau",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Point d’eau supprimé avec succès"
                    },
                    "404": {
                        "description": "Point d’eau non trouvé"
                    }
                }
            }
        },
...
        "/api/user.delete/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a User",
                "description": "Delete a user by their ID.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}